Updated writer and source time function
