+++ include/source/moment_tensor_source.hpp
@@ -5,7 +5,7 @@
 #include "kokkos_abstractions.h"
 #include "quadrature/interface.hpp"
 #include "source.hpp"
-#include "source_time_function.h"
+#include "source_time_function/interface.hpp"
 #include "specfem_mpi.h"
 #include "specfem_setup.hpp"
 #include "utils.h"
