+++ include/source/source.hpp
@@ -4,7 +4,7 @@
 #include "enums.h"
 #include "kokkos_abstractions.h"
 #include "quadrature/interface.hpp"
-#include "source_time_function.h"
+#include "source_time_function/interface.hpp"
 #include "specfem_mpi.h"
 #include "specfem_setup.hpp"
 #include "utils.h"
