+++ include/source_time_function/interface.hpp
@@ -1,0 +1,7 @@
+#ifndef _SOURCE_TIME_FUNCTION_INTERFACE_HPP
+#define _SOURCE_TIME_FUNCTION_INTERFACE_HPP
+
+#include "dirac.hpp"
+#include "source_time_function.hpp"
+
+#endif
