+++ include/writer/interface.hpp
@@ -1,0 +1,7 @@
+#ifndef _WRITER_INTERFACE_HPP
+#define _WRITER_INTERFACE_HPP
+
+#include "seismogram.hpp"
+#include "writer.hpp"
+
+#endif
