+++ src/source/moment_tensor_source.cpp
@@ -3,7 +3,7 @@
 #include "kokkos_abstractions.h"
 #include "quadrature/interface.hpp"
 #include "source/interface.hpp"
-#include "source_time_function.h"
+#include "source_time_function/interface.hpp"
 #include "specfem_mpi.h"
 #include "specfem_setup.hpp"
 #include "utilities.cpp"
