+++ src/writer/seismogram.cpp
@@ -1,6 +1,6 @@
-#include "writer.h"
 #include "compute/interface.hpp"
 #include "receiver/interface.hpp"
+#include "writer/interface.hpp"
 #include <fstream>
 
 void specfem::writer::seismogram::write() {
